==> R-scripts/2_data_prep.Rmd <==
---
output: github_document
---

In this script, the data is processed and formatted to be used in the EDA and model fit.  

Import libraries 
```{r}
library(dplyr)
library(tidyr)
library(ggplot2)
theme_set(theme_bw())
```

Load in data
```{r}
#Ideal
load("../R-scripts/Rdata/alleles_capt_ideal_twopop.Rdata")
#Realistic
#load("../R-scripts/Rdata/alleles_capt_realistic_twopop.Rdata")
```
***

### GET DATA IN PROPER FORMAT FOR PLOTTING AN ANALYSIS
We need all the data which is currently in 3 separate 3D matrices to be combined into one large matrix in tidy format

1. Converting 3D results matrices to 2D for plotting and analysis purposes
```{r}
same_long=NULL
eligible_long=NULL
skewed_long=NULL
j=1
for(i in 1:50) {
    temp = rbind(prop_capt_all_same[,,j], prop_capt_all_same[,,(j+1)])
    same_long = rbind(same_long, temp)
    j = j+2
    if(j>50) {break}
}
j=1
for(i in 1:50) {
    temp = rbind(prop_capt_all_eligible[,,j], prop_capt_all_eligible[,,(j+1)])
    eligible_long = rbind(eligible_long, temp)
    j = j+2
    if(j>50) {break}
}
j=1
for(i in 1:50) {
    temp = rbind(prop_capt_skewed[,,j], prop_capt_skewed[,,(j+1)])
    skewed_long = rbind(skewed_long, temp)
    j = j+2
    if(j>50) {break}
}
rm(temp)
```
2. Binding all data frames together into one huge data frame 
This is considered tidy format? 
```{r}
tidy_df = rbind(same_long, eligible_long, skewed_long)
tidy_df = as.data.frame(tidy_df)
#' Just adding a column indicating the number of failures (number of alleles not captured) for binomial data
tidy_df$num_fails = c(as.numeric(tidy_df$total_alleles) - as.numeric(tidy_df$num_capt))
save(tidy_df, file="../../R-scripts/Rdata/tidy_df_realistic_twopop.Rdata")
```
